Updated project .net framework version from 4.0 to 3.5
Yes - you read that right.  I needed to go back in time a bit...needed
to downgrade the .net framework used by this project in order to get
parallel test execution to work with PNUnit.  This also forced me to
updated my references to point to the .net 3.5 versions of the Webdriver
DLLs.  Also cleaned up some unreferenced classes and methods in the
process.